refactor : 아웃바운드 어댑터 디렉토리의 상위 디렉토리 추가 (#37)
